ft - remove the dashboard activity and refactor mealsearchactivity
