+++ app/src/main/java/com/chachaup/chachaup/ui/MainActivity.java
@@ -1,86 +1,109 @@
 package com.chachaup.chachaup.ui;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.chachaup.chachaup.Constants;
 import com.chachaup.chachaup.R;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class MainActivity extends AppCompatActivity{
-    @BindView(R.id.buttonLogin)
-    Button mLoginButton;
-    @BindView(R.id.editTextEmailLogin)
-    EditText mEmail;
-    @BindView(R.id.editTextPasswordLogin)
-    EditText mPassword;
-    @BindView(R.id.textViewSignUp)
-    TextView mSignUp;
-    @BindView(R.id.textViewWarning)
-    TextView mWarning;
+public class MealSearchActivity extends AppCompatActivity implements View.OnClickListener {
+    @BindView(R.id.buttonFindMeal)
+    Button mFindMeal;
+    @BindView(R.id.buttonSavedMeals) Button mSavedMeals;
 
-//    private boolean isAllFieldsChecked;
+    private FirebaseAuth mAuth;
+    private FirebaseAuth.AuthStateListener mAuthListener;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_search_meal);
         ButterKnife.bind(this);
 
-//        isAllFieldsChecked = CheckAllFields();
+        mFindMeal.setOnClickListener(this);
+        mSavedMeals.setOnClickListener(this);
 
-        mSignUp.setOnClickListener(new View.OnClickListener() {
+        mAuth = FirebaseAuth.getInstance();
+        mAuthListener = new FirebaseAuth.AuthStateListener() {
             @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this,SignUp.class);
-                Toast.makeText(MainActivity.this,"Taking you to sign up...",Toast.LENGTH_SHORT).show();
-                startActivity(intent);
-            }
-        });
-
-        mLoginButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String userEmail = mEmail.getText().toString();
-                String userPassword = mPassword.getText().toString();
-                Intent intent = getIntent();
-                String email = intent.getStringExtra("email");
-                String pass = intent.getStringExtra("pass");
-                String name = intent.getStringExtra("name");
-
-                if(userEmail.equals(email) && userPassword.equals(pass)){
-                    Toast.makeText(MainActivity.this,"Signing in...", Toast.LENGTH_SHORT).show();
-                    Intent intent1 = new Intent(MainActivity.this, MealSearchActivity.class);
-                    intent1.putExtra("name",name);
-                    startActivity(intent1);
-                } else{
-                    mWarning.setText(getString(R.string.warning));
+            public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
+                FirebaseUser user = firebaseAuth.getCurrentUser();
+                if (user != null){
+                    getSupportActionBar().setTitle("Welcome " + user.getDisplayName() + "!");
                 }
-
             }
-        });
+        };
     }
-    // validation of fields
-//    private boolean CheckAllFields() {
-//        if(mEmail.length() == 0){
-//            mEmail.setError("This field is required");
-//            return false;
-//        }
-//        if(mPassword.length() == 0){
-//            mPassword.setError("This field is required");
-//            return false;
-//        } else if (mPassword.length() < 8){
-//            mPassword.setError("Password must be a minimum of 8 characters");
-//            return false;
-//        }
-//        return true;
-//    }
+    @Override
+    public void onClick(View v) {
+        if (v == mFindMeal){
+            Intent intent = new Intent(MealSearchActivity.this, MealsActivity.class);
+            Toast.makeText(MealSearchActivity.this,"Searching meal...",Toast.LENGTH_SHORT).show();
+            startActivity(intent);
+        }
+
+        if (v == mSavedMeals){
+            Intent intent = new Intent(MealSearchActivity.this, SavedRecipeListActivity.class);
+            startActivity(intent);
+        }
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.menu_main, menu);
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        int id = item.getItemId();
+        if (id == R.id.actionLogOut){
+            logout();
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
+    private void logout(){
+        FirebaseAuth.getInstance().signOut();
+        Intent intent = new Intent(MealSearchActivity.this, LoginActivity.class);
+        intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+        startActivity(intent);
+        finish();
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        mAuth.addAuthStateListener(mAuthListener);
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        mAuth.removeAuthStateListener(mAuthListener);
+    }
 }
